rbapi/requirements.txt
Django==3.0.2
django-jet==1.0.8
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
reportlab==3.5.34